Support closure handlers (#1) (Doesn't support return values from closures yet
Implemented by creating a c++ lambda which call the nim closure

Add support for calling callbacks in nim code

+++ src/slinim.nim
@@ -61,6 +61,11 @@ when isMainModule:
     let outputFile = getTempDir() / "slinim_" & $SecureHash(hash.finalize()) & ".nim"
     let nimFile = slintFile.replace(".slint", ".nim")
     outputFile.writeFile(bindings)
-    nimFile.writeFile(fmt"""include "{outputFile}" """)
+    # Check we aren't overwriting anything important first
+    if not nimFile.fileExists() or nimFile.getFileInfo().size <= 75:
+      nimFile.writeFile(fmt"""include "{outputFile}" """)
+    else:
+      stderr.writeLine("Output file already exists")
+      quit 1
   else:
     outputFile.writeFile(bindings)
